Allow task files to have frontmatter
"Frontmatter" is a delimited block at the start of a file, often used
to add metadata to Markdown files by some static site publishers. Next
expects the frontmatter delimiters to be triple backticks (```).

Ensure anything that would match a todo in the frontmatter is not
erroneously found as a todo.